started a.h file just to get the ball rolling
Here is a .h file that was made with the intention of creating a version of the program with a linked list or a vector to store answers.  Also included in this upload a possible format for the .txt file.  I realize now how unclean it looks, I've sort of piece it together 2 mins here 2 mins there.  Will get a good chunk of time in tomorrow.