updated header for vectors
So I've gone ahead and worked on this .h file, making it vector centric.  I've also separated the program into two stages.  The first stage is everything that is necessary to play the 20 questions game.  The second stage is primarily adding answers and questions.  Stage 2 also has a method for printing all possible answers