WORKING 20 QUESTION GAME
Right, so, it does not have all the capability that we talked about.  It can't add a new questions, or answers.  It is NOT smart.

It reads in a file full of questions and answers with true/false for each question.  It lets you play the 20 questions game, there are also a number of print options in the print function.  I was printing all sorts of different things and so there is a lot of commented out code in that print function.

There is still a ton to do and improve, but it works and could be turned in